refactor(modularization): Phase 3 Step 5a — KB routes move to domain + AuthenticatedUser Protocol + MarketplaceOps sharing methods + 7 route schemas carve-out (#16)
Step 5a in the Phase 3 #1629 atomic sequence — the final sub-step
that completes the Phase 3 KB domain shape (per @符炫炜 msg=99183dff
5c → 5b → 5a sub-ordering).

**routes move (aperag/views/{collections,documents}_v2.py → aperag/domains/knowledge_base/api/routes.py)**:
- The two legacy view modules are merged into a single canonical
  ``APIRouter`` at ``aperag/domains/knowledge_base/api/routes.py``.
  Handlers are byte-identical in URL, method, response_model,
  audit-decorator, status codes — only the parameter types for
  ``Depends(required_user)`` change from ``User`` ORM to the new
  ``AuthenticatedUser(Protocol)``.
- ``aperag/views/collections_v2.py`` and ``aperag/views/documents_v2.py``
  collapse to router-import shims re-exporting the same
  ``APIRouter`` object, so pre-migration callers keep working.
- ``aperag/app.py`` replaces the two ``*_router`` includes with a
  single ``knowledge_base_router`` include (same ``/api/v2`` prefix),
  reducing the include list by one line.

**AuthenticatedUser(Protocol) in ports.py**:
- Per-domain auth context (lesson 9a-ter, Phase 4 canonical
  microadjust msg=d47fa490 Section 3). Only pins the user ``id``
  attribute KB routes actually read. Satisfies the
  ``test_aperag_domains_auth_dependency_is_not_any`` boundary gate
  for the moved route handlers.

**MarketplaceOps Protocol extended**:
- Added ``get_sharing_status`` / ``publish_collection`` /
  ``unpublish_collection`` methods so the
  ``/collections/{id}/sharing`` group can reach marketplace through
  the DI accessor instead of direct ``marketplace_service`` import.
  Concrete legacy ``marketplace_service`` already provides all three
  methods → structurally satisfies the extended Protocol with zero
  provider changes.

**7 route schemas carve-out (kb.schemas + view_models)**:
- ``ConfirmDocumentsRequest`` / ``FetchUrlRequest`` /
  ``DeleteDocumentsRequest`` / ``DeleteDocumentsResponse`` /
  ``SharingStatusResponse`` / ``MineruTokenTestRequest`` /
  ``MineruTokenTestResponse`` moved from
  ``aperag.schema.view_models`` to
  ``aperag.domains.knowledge_base.schemas``. All are KB-facing
  route input/output envelopes; the dual-hook reexport keeps
  pre-migration callers transparent.

**Test update**:
- ``tests/unit_test/test_collection_v2_openapi_contract.py::test_collection_v2_does_not_occupy_documents_or_search_or_graph``
  → renamed ``test_collection_v2_does_not_occupy_search_or_graph``
  and the ``/documents`` exclusion dropped, because Step 5a
  intentionally merges collections_v2 + documents_v2 into a single
  KB router. Retrieval ``/searches`` and knowledge_graph
  ``/graphs`` exclusions remain — those live in other domain
  routers.

**FastAPI note**:
- The moved routes module intentionally does NOT use
  ``from __future__ import annotations``. FastAPI inspects return
  type annotations at route-registration time to decide whether a
  status-204 handler has a response body; under PEP 563 annotations
  become strings and the ``-> Response`` return annotation fails
  the ``is_body_allowed_for_status_code(204)`` assertion. Dropping
  the future import keeps the 204 command routes green.

**Verification**:
- ``scripts/export_openapi.py --check`` clean — OpenAPI spec
  byte-stable across the router move.
- Full unit suite green: 669 passed / 29 skipped. G1 AST walk green
  (the moved ``api/routes.py`` is under ``aperag/domains/**`` and
  imports zero legacy aggregate modules).
- ``uvx ruff check`` + ``uvx ruff format --check`` clean across 271
  ``aperag/`` files.
- Handler routes are structurally identical: 24 routes on the
  merged KB router, each with the same URL + method + status code
  + response model as the pre-move code.

Phase 3 Steps 2 → 5a sequence is now complete in this PR head;
Steps 6 / 7 / 8 (MCP update, new boundary tests, docs/breaking-
changes, full gate sweep + self-merge) remain.
